changed type of all instances of fd in protocol.h to int64_t.
* changed type of fd to int64_t from uint64_t
  - fd = -1 is used as indication of not passing a valid fd, for xattrop and gf_file_lk calls.

* made relevant changes in protocol/{client,server} and fd.{c,h}

+++ libglusterfs/src/fd.c
@@ -271,7 +271,7 @@ gf_fd_put (fdtable_t *fdtable, int32_t fd)
 
 
 fd_t *
-gf_fd_fdptr_get (fdtable_t *fdtable, int32_t fd)
+gf_fd_fdptr_get (fdtable_t *fdtable, int64_t fd)
 {
 	fd_t *fdptr = NULL;
